Update README.md
Packed with all important micronutrients, and with a balanced profile of macronutrients, Strawberries are considered to be some of the best fruits out there, they are sodium free, fat free, cholesterol free, and low in sugar, high in fibre, vitamins, and antioxidants, there have never been any fruit that can match their benefit and taste. This repository is meant to create a list of mathematical functions that are derived from Strawberry benefits for our general health, but the main goal is to create different models that can mimic the nutrition benefits of strawberries into real life applications and examples in different aspects.